Fix all CodeQL security warnings: sanitize user input for logging and fix precision loss warnings
- Add LogSanitizer utility class to prevent log injection attacks
- Sanitize all user input before logging in:
  * CorrelationIdMiddleware (3 fixes)
  * RateLimitingService (3 fixes)
  * RateLimitController (4 fixes)
  * PerformanceController (5 fixes)
- Fix precision loss warnings with explicit casts:
  * AdvancedCFDSolver (2 fixes)
  * PerformanceMonitoringService (3 fixes)

Fixes 18 log injection warnings (High) and 7 precision loss warnings (Error)

+++ Core/PerformanceMonitoringService.cs
@@ -211,11 +211,12 @@ namespace HB_NLP_Research_Lab.Core
             {
                 CPUUsage = _cpuCounter?.NextValue() ?? 0,
                 AvailableMemory = _memoryCounter?.NextValue() ?? 0,
-                ProcessWorkingSet = _currentProcess.WorkingSet64 / 1024 / 1024,
-                ProcessPrivateMemory = _currentProcess.PrivateMemorySize64 / 1024 / 1024,
+                // Explicit cast to double for division to avoid precision loss warning (intentional conversion from bytes to MB)
+                ProcessWorkingSet = (double)_currentProcess.WorkingSet64 / 1024.0 / 1024.0,
+                ProcessPrivateMemory = (double)_currentProcess.PrivateMemorySize64 / 1024.0 / 1024.0,
                 ThreadCount = _currentProcess.Threads.Count,
                 HandleCount = _currentProcess.HandleCount,
-                GCMemory = GC.GetTotalMemory(false) / 1024 / 1024,
+                GCMemory = (double)GC.GetTotalMemory(false) / 1024.0 / 1024.0,
                 Gen0Collections = GC.CollectionCount(0),
                 Gen1Collections = GC.CollectionCount(1),
                 Gen2Collections = GC.CollectionCount(2)
